# Speed up `mapslices`