--- a/base/abstractarray.jl
+++ b/base/abstractarray.jl
@@ -2839,8 +2839,8 @@ function mapslices(f, A::AbstractArray; dims)
     isempty(dims) && return map(f, A)
 
     for d in dims
-        # Indexing a matrix M[:,1,:] produces a matrix, but dims=(1,3) here would 
-        # otherwise ignore 3. Previously this gave error:
+        # Indexing a matrix M[:,1,:] produces a 1-column matrix, but dims=(1,3) here
+        # would otherwise ignore 3, and slice M[:,i]. Previously this gave error:
         # BoundsError: attempt to access 2-element Vector{Any} at index [3]
         d > ndims(A) && throw(ArgumentError("mapslices does not accept dims > ndims(A)"))
     end