This does a little tidying up of `mapslices`, to work internally with tuples instead of mutating arrays of indices, and to avoid broadcasting in the inner loop where possible. There should be no functional changes. It's still not type stable, but it is a bit faster:
```julia
julia> @btime mapslices(cumsum, $(rand(10,100)), dims=1);
  26.250 μs (631 allocations: 39.47 KiB)  # master
  4.756 μs (116 allocations: 22.45 KiB))  # this PR

julia> @btime reduce(hcat, map(cumsum, eachcol($(rand(10,100)))));  # perhaps preferred?
  2.898 μs (102 allocations: 22.88 KiB)

julia> @btime cumsum($(rand(10,100)), dims=1); # for reference, without slices
  409.091 ns (1 allocation: 7.94 KiB)

julia> @btime mapslices(prod, $(rand(100,10)), dims=2);
  25.500 μs (835 allocations: 23.25 KiB)  # master
  1.433 μs (21 allocations: 1.50 KiB)     # this PR

julia> @btime map(prod, eachrow($(rand(100,10))));  # certainly better!
  526.393 ns (1 allocation: 896 bytes)

julia> @btime prod($(rand(100,10)), dims=2); # for reference, without slices
  292.286 ns (5 allocations: 976 bytes)
```