@mcabbott 2022-02-09
Pre-1.8 bump.

This still hits https://github.com/JuliaLang/julia/issues/43064, but IIRC reverting https://github.com/JuliaLang/julia/pull/40996/commits/90cc0277fcf35dcabefd7be8f950df3d0d227e2c would be enough to avoid that, at the cost of losing type stability.