@mcabbott 2022-02-09
BTW, the same benchmarks as the top message, on the same computer, are significantly slower on today's master. I hope this is tracked elsewhere? The relative change of the PR is still good:
```julia
julia> @btime mapslices(cumsum, $(rand(10,100)), dims=1);
  min 55.250 μs, mean 60.423 μs (630 allocations, 37.62 KiB)  # master
  min 8.111 μs, mean 9.557 μs (114 allocations, 22.45 KiB)  # this PR
  
julia> @btime reduce(hcat, map(cumsum, eachcol($(rand(10,100)))));  # perhaps preferred?
  min 4.875 μs, mean 6.207 μs (102 allocations, 22.88 KiB)

julia> @btime cumsum($(rand(10,100)), dims=1); # for reference, without slices
  min 699.885 ns, mean 1.865 μs (1 allocation, 7.94 KiB)

julia> @btime mapslices(prod, $(rand(100,10)), dims=2);
  min 57.458 μs, mean 61.089 μs (834 allocations, 22.91 KiB)  # master
  min 2.787 μs, mean 2.931 μs (18 allocations, 1.48 KiB)  # this PR

julia> @btime map(prod, eachrow($(rand(100,10))));  # certainly better!
  min 938.259 ns, mean 978.073 ns (1 allocation, 896 bytes)

julia> @btime prod($(rand(100,10)), dims=2); # for reference, without slices
  min 510.854 ns, mean 553.738 ns (5 allocations, 976 bytes)
```