@mcabbott 2022-05-07
Bump?

This is nearly a year old, and seems an uncontroversial improvement. 

At present it hits https://github.com/JuliaLang/julia/issues/43064, I can remove one commit (and perhaps leave a comment) to avoid that, at the cost of losing type-stability. But still a large speedup.